[clang][CodeGen] Avoid emitting unnecessary memcpy of records without content
Clang codegen emits an unnecessary memcpy for C++ class without content. This patch fixes the issue.

See https://github.com/llvm/llvm-project/issues/59336

Please let me know if the patch is going into the right direction that it can be accepted. If yes, I'll finish it with test cases and also extend it to C.